+++ Forge/src/main/java/fuzs/puzzleslib/client/core/ForgeClientFactories.java
@@ -1,5 +1,7 @@
 package fuzs.puzzleslib.client.core;
 
+import fuzs.puzzleslib.core.ContentRegistrationFlags;
+
 import java.util.function.Consumer;
 
 /**
@@ -8,7 +10,7 @@ import java.util.function.Consumer;
 public final class ForgeClientFactories implements ClientFactories {
 
     @Override
-    public Consumer<ClientModConstructor> clientModConstructor(String modId) {
-        return constructor -> ForgeClientModConstructor.construct(modId, constructor);
+    public Consumer<ClientModConstructor> clientModConstructor(String modId, ContentRegistrationFlags... contentRegistrations) {
+        return constructor -> ForgeClientModConstructor.construct(constructor, modId, contentRegistrations);
     }
 }