+++ Forge/src/main/java/fuzs/puzzleslib/impl/biome/BiomeLoadingHandler.java
@@ -34,11 +34,11 @@ public class BiomeLoadingHandler {
 
     public static void register(String modId, IEventBus modEventBus, Multimap<BiomeLoadingPhase, BiomeModificationData> biomeLoadingEntries) {
         DeferredRegister<BiomeModifier> biomeModifiersRegistry = DeferredRegister.create(ForgeRegistries.Keys.BIOME_MODIFIERS, modId);
-        DeferredRegister<Codec<? extends BiomeModifier>> biomeModifierSerializersRegitry = DeferredRegister.create(ForgeRegistries.Keys.BIOME_MODIFIER_SERIALIZERS, modId);
+        DeferredRegister<Codec<? extends BiomeModifier>> biomeModifierSerializersRegistry = DeferredRegister.create(ForgeRegistries.Keys.BIOME_MODIFIER_SERIALIZERS, modId);
         biomeModifiersRegistry.register(modEventBus);
-        biomeModifierSerializersRegitry.register(modEventBus);
+        biomeModifierSerializersRegistry.register(modEventBus);
         BiomeModifierImpl biomeModifier = new BiomeModifierImpl(biomeLoadingEntries);
-        biomeModifierSerializersRegitry.register("biome_modifiers_codec", () -> biomeModifier.codec());
+        biomeModifierSerializersRegistry.register("biome_modifiers_codec", () -> biomeModifier.codec());
         biomeModifiersRegistry.register("biome_modifiers", () -> biomeModifier);
     }